added HW3 files

+++ lab1/test.c
@@ -2,5 +2,5 @@
 
 int main()
 {
-	printf( "%d\n", sizeof(long));
+	printf( "%d\n", (int)sizeof(long));
 }
